dfdf

+++ sefafx.cpp
@@ -1,23 +1,23 @@
 #include "sefafx.h"
 
-int SefAxf::GetFileSize(const char* pszPath_)
+long int SefAxf::GetFileSize(const char* pszPath_)
 {
-  struct stat _sBuf;
-  if (stat(pszPath_, &_sBuf))
-  {
-    return -1;
-  }
-  
-  return (int)_sBuf.st_size;
+    struct stat _sBuf;
+    if (stat(pszPath_, &_sBuf))
+    {
+        return -1;
+    }
+
+    return (long int)_sBuf.st_size;
 }
 
-int SefAxf::GetFileSize(int fd_)
+long int SefAxf::GetFileSize(int fd_)
 {
-  struct stat _sBuf;
-  if (fstat(fd_, &_sBuf))
-  {
-    return -1;
-  }
-  
-  return (int)_sBuf.st_size;
+    struct stat _sBuf;
+    if (fstat(fd_, &_sBuf))
+    {
+        return -1;
+    }
+
+    return (long int)_sBuf.st_size;
 }
